#Conflicts:
#	Cargo.lock
#	api/rs/slint/tests/partial_renderer.rs
#	demos/Cargo.lock
#	examples/Cargo.lock
#	internal/backends/winit/Cargo.toml
#	internal/backends/winit/winitwindowadapter.rs
#	internal/compiler/builtins.slint
#	internal/compiler/generator/cpp.rs
#	internal/compiler/generator/rust.rs
#	internal/compiler/llr/optim_passes/remove_unused.rs
#	internal/compiler/typeregister.rs
#	internal/core/window.rs
#	internal/interpreter/dynamic_item_tree.rs
#	internal/interpreter/eval.rs
#	tests/Cargo.lock
#	tests/backends/tests/cases/mod.rs

+++ examples/plotter/main.rs
@@ -17,8 +17,8 @@ slint::slint! {
 fn pdf(x: f64, y: f64, a: f64) -> f64 {
     const SDX: f64 = 0.1;
     const SDY: f64 = 0.1;
-    let x = x as f64 / 10.0;
-    let y = y as f64 / 10.0;
+    let x = x / 10.0;
+    let y = y / 10.0;
     a * (-x * x / 2.0 / SDX / SDX - y * y / 2.0 / SDY / SDY).exp()
 }
